🐛 replace BigInt-forbidden types and stop auto-overwriting bindings.ts
Specta refuses to export usize/i64/u64 fields to TypeScript (precision loss
risk), which crashed the app on startup: `tauri dev` calls .export(...) at
runtime, and every M1 struct built during the session used usize for counts/
indices and i64/u64 for timestamps. Switched counts/indices (offset, limit,
col, color_index, additions, deletions, total_count) to u32, and Unix
timestamps (timestamp_sec, author_timestamp_sec, timestamp_ms,
last_opened_at_ms) to f64 — safe up to 2^53, far beyond any real epoch value.

Also removed the .export(...) call in lib.rs entirely: it was overwriting
the hand-written src/ipc/bindings.ts with tauri-specta's own generated
shape (a `commands` object returning {status, data|error} instead of the
Promise<T> that src/ipc/client.ts's invokeCommand wrappers and their
browser-dev mocks assume), which would have broken every IPC call site.

+++ src-tauri/src/commands/app.rs
@@ -45,7 +45,7 @@ pub fn simulate_repo_change(app: tauri::AppHandle, repo_path: String) -> Result<
     let now = SystemTime::now()
         .duration_since(UNIX_EPOCH)
         .unwrap_or_default()
-        .as_millis() as u64;
+        .as_millis() as f64;
 
     let payload = RepoChangedPayload {
         repo_path,
